[12128] Fix more warnings - close #4
* Remove warning: enumeration value '<value>' not handled in switch
* Remove warning: enumeral and non-enumeral type in conditional expression
* Remove warning: too many arguments for format

+++ src/game/BattleGround/BattleGround.cpp
@@ -73,7 +73,7 @@ namespace MaNGOS
                 data << ObjectGuid(targetGuid);
                 data << uint32(strlen(text) + 1);
                 data << text;
-                data << uint8(i_source ? i_source->GetChatTag() : CHAT_TAG_NONE);
+                data << uint8(i_source ? i_source->GetChatTag() : uint8(CHAT_TAG_NONE));
             }
 
             ChatMsg i_msgtype;
@@ -152,7 +152,7 @@ namespace MaNGOS
                 data << ObjectGuid(targetGuid);
                 data << uint32(strlen(str) + 1);
                 data << str;
-                data << uint8(i_source ? i_source->GetChatTag() : CHAT_TAG_NONE);
+                data << uint8(i_source ? i_source->GetChatTag() : uint8(CHAT_TAG_NONE));
             }
         private:
