+++ src/game/Channel.cpp
@@ -574,7 +574,7 @@ void Channel::Say(ObjectGuid p, const char* what, uint32 lang)
         data << ObjectGuid(p);
         data << uint32(messageLength);
         data << what;
-        data << uint8(plr ? plr->GetChatTag() : CHAT_TAG_NONE);
+        data << uint8(plr ? plr->GetChatTag() : uint8(CHAT_TAG_NONE));
 
         SendToAll(&data, !m_players[p].IsModerator() ? p : ObjectGuid());
     }
